// CPSNext/Jha.Cps.Next.Web.Angular/apps/cps-next-ui/src/app/cpsNext-Legacy/entities/enums.ts
export enum LOG_LEVELS {
  ALL = 0,
  DEBUG = 1,
  INFO = 2,
  WARN = 3,
  ERROR = 4,
  FATAL = 5,
  OFF = 6
}


export enum PRODUCT_NAME {
  Introduction = 1,
  CardMaintenance = 2,
  CPSalerts = 3,
  IPay = 4
}


export enum ERROR_REASON {
  Exception = 1,
  NotAuthorized = 2,
  SesssinExpired = 3,
  NullUserContext = 4
}